+++ pages/line/@id/+data.ts
@@ -4,7 +4,7 @@ import { LineCollection } from "@/server/data/line/line-collection";
 import { Line } from "@/server/data/line/line";
 import { JsonSerializable } from "@/shared/json-serializable";
 import { CalendarData } from "@/shared/types/calendar-data";
-import { createCalendarData } from "@/server/data/disruption/period/utils/render-calendar-marks";
+import { createCalendarData } from "@/server/data/disruption/period/utils/create-calendar-data";
 import { EndsExactly } from "@/server/data/disruption/period/ends/ends-exactly";
 import { EveningsOnlyDisruptionPeriod } from "@/server/data/disruption/period/evenings-only-disruption-period";
 import { StandardDisruptionPeriod } from "@/server/data/disruption/period/standard-disruption-period";
